+++ mollib/src/main/java/com/bammellab/mollib/objects/SegmentAtomToAtomBond.kt
@@ -1,17 +1,14 @@
 /*
- * Copyright (C) 2016-2018 James Andreas
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
+ *  Copyright 2020 Bammellab / James Andreas
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License
  */
 
 @file:Suppress("unused")
@@ -19,6 +16,7 @@
 package com.bammellab.mollib.objects
 
 import com.bammellab.mollib.common.math.MathUtil
+import com.bammellab.mollib.common.math.MathUtil.PIFLOAT
 import com.bammellab.mollib.common.math.MotmVector3
 import com.bammellab.mollib.objects.GlobalObject.BRIGHTNESS_FACTOR
 import com.kotmol.pdbParser.AtomInformationTable
@@ -49,9 +47,9 @@ class SegmentAtomToAtomBond(private val molecule: Molecule) {
         val positionEnd = MotmVector3(atom2.atomPosition)
         val positionMid = MotmVector3()
 
-        positionMid.x = (positionEnd.x + positionStart.x) / 2.0
-        positionMid.y = (positionEnd.y + positionStart.y) / 2.0
-        positionMid.z = (positionEnd.z + positionStart.z) / 2.0
+        positionMid.x = (positionEnd.x + positionStart.x) / 2.0f
+        positionMid.y = (positionEnd.y + positionStart.y) / 2.0f
+        positionMid.z = (positionEnd.z + positionStart.z) / 2.0f
 
         var elementSymbol: String = atom1.elementSymbol
         var ai = AtomInformationTable.atomSymboltoAtomNumNameColor[elementSymbol]
@@ -76,7 +74,7 @@ class SegmentAtomToAtomBond(private val molecule: Molecule) {
         //        MotmVector3 p1p2 = new MotmVector3();
         //        p1p2.setAll(position_end);
         //        p1p2.subtract(position_start);
-        //        MotmVector3 P = new MotmVector3(Math.random(), Math.random(), Math.random());
+        //        MotmVector3 P = new MotmVector3(Math.random().toFloat(), Math.random().toFloat(), Math.random());
         //        MotmVector3 R = new MotmVector3(p1p2);
         //        R.cross(P);
         //        MotmVector3 S = new MotmVector3(R);
@@ -86,9 +84,9 @@ class SegmentAtomToAtomBond(private val molecule: Molecule) {
 
         p1p2.setAll(positionEnd)
         p1p2.subtract(positionStart)
-        pvec.x = Math.random()
-        pvec.y = Math.random()
-        pvec.z = Math.random()
+        pvec.x = Math.random().toFloat()
+        pvec.y = Math.random().toFloat()
+        pvec.z = Math.random().toFloat()
         // MotmVector3 R = new MotmVector3(p1p2);
         rvec.setAll(p1p2)
         rvec.cross(pvec)
@@ -101,12 +99,12 @@ class SegmentAtomToAtomBond(private val molecule: Molecule) {
         rvec.normalize()
         svec.normalize()
 
-        var x1: Double
-        var y1: Double
-        var z1: Double
-        var x2: Double
-        var y2: Double
-        var z2: Double
+        var x1: Float
+        var y1: Float
+        var z1: Float
+        var x2: Float
+        var y2: Float
+        var z2: Float
 
         /*
          * math: two tris per slice, wrapping for numSlices+1 (hit original again to close)
@@ -128,10 +126,10 @@ class SegmentAtomToAtomBond(private val molecule: Molecule) {
         var n: FloatArray
         while (i <= numSlices) {
 
-            val angleInRadians1 = (i.toDouble() / numSlices.toDouble()
-                    * Math.PI * 2.0)
-            val angleInRadians2 = ((i + 1).toDouble() / numSlices.toDouble()
-                    * Math.PI * 2.0)
+            val angleInRadians1 = (i.toFloat() / numSlices.toFloat()
+                    * PIFLOAT * 2.0f)
+            val angleInRadians2 = ((i + 1).toFloat() / numSlices.toFloat()
+                    * PIFLOAT * 2.0f)
 
             val s1 = radius * MathUtil.sin(angleInRadians1)
             val s2 = radius * MathUtil.sin(angleInRadians2)
